capta-api/requirements.txt
aniso8601==9.0.1
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
pandas==1.3.0
pycparser==2.20
PyJWT==2.1.0
pyodbc==4.0.31
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1